Remove jsx extension
