+++ frontend/components/OAuth/oauth.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react';
+import { Text, View } from 'react-native';
+import FBSDK from 'react-native-fbsdk';
+const LoginButton = FBSDK.LoginButton;
+const AccessToken = FBSDK.AccessToken;
+class OAuth extends Component {
+  render() {
+    return (
+      <View></View>
+    );
+  }
+}
+
+export default OAuth;
