--- frontend/components/OAuth/oauth.jsx
@@ -1,39 +1,0 @@
-import React, { Component } from 'react';
-import { Text, View } from 'react-native';
-import FBSDK from 'react-native-fbsdk';
-//
-// const {
-//   LoginButton,
-//   AccessToken
-// } = FBSDK;
-
-const LoginButton = FBSDK.LoginButton;
-const AccessToken = FBSDK.AccessToken;
-class OAuth extends Component {
-  render() {
-    return (
-  <View></View>
-    // <LoginButton
-    //   publishPermissions={["publish_actions"]}
-    //   onLoginFinished={
-    //     (error, result) => {
-    //       if (error) {
-    //         alert("login has error: " + result.error);
-    //       } else if (result.isCancelled) {
-    //         alert("login is cancelled.");
-    //       } else {
-    //         AccessToken.getCurrentAccessToken().then(
-    //           (data) => {
-    //             alert(data.accessToken.toString());
-    //           }
-    //         );
-    //       }
-    //     }
-    //   }
-    //   onLogoutFinished={() => alert("logout.")}/>
-
-);
-  }
-}
-
-export default OAuth;
