+++ frontend/components/OAuth/oauth_container.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { signup } from '../../actions/session_actions';
-import OAuth from './oauth';
+import OAuth from './oauth.js';
 
 const mapDispatchToProps = dispatch => ({
   signup: (user) => dispatch(signup(user)),
